studymind-ai/requirements.txt
streamlit==1.35.0
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-text-splitters==0.2.2
openai==1.35.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1
PyMuPDF==1.24.5
pdfplumber==0.11.1
python-docx==1.1.2
python-pptx==0.6.23
networkx==3.3
pyvis==0.3.2
python-dotenv==1.0.1
bcrypt==4.1.3
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
gTTS==2.5.1
SpeechRecognition==3.10.4
fpdf2==2.7.9
keybert==0.7.0